LangChain_Demo/pyproject.toml
[project]
name = "langchain-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "huggingface-hub>=1.7.1",
    "langchain>=1.2.12",
    "langchain-anthropic>=1.3.5",
    "langchain-core>=1.2.19",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
